Split Passes.scala into separate files
